requirements.txt
python-dotenv
google-generativeai
tqdm
loguru
ratelimit
tenacity